requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.4
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
black==24.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
docstring-parser==0.15
executing==2.0.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
iniconfig==2.0.0
inline-snapshot==0.8.0
instructor==1.4.3
datasets==2.21.0
langsmith==0.1.48
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.40.1
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pydantic==2.7.0
pydantic-core==2.18.1
pygments==2.17.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-asyncio-cooperative==0.36.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
tokenizers==0.19.1
toml==0.10.2
tqdm==4.66.2
typer==0.12.3
types-toml==0.10.8.20240310
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
boto3==1.35.29
autoevals==0.0.87
braintrust==0.0.155
google-generativeai==0.5.4
